Refactor AdsDao
