+++ src/main/java/io/bittiger/ads/web/SearchResult.java
@@ -3,6 +3,7 @@ package io.bittiger.ads.web;
 import com.google.gson.Gson;
 import io.bittiger.ads.AdsEngine;
 import io.bittiger.ads.activity.*;
+import io.bittiger.ads.datastore.AdsDao;
 import io.bittiger.ads.util.Ad;
 import io.bittiger.ads.util.AllocationType;
 import org.json.JSONArray;
